refactor: More Tab to Account 로 변경
