+++ presentation/src/main/java/wolf/shin/earlyadopter/model/screen_router/AccountScreenRouter.kt
@@ -1,0 +1,6 @@
+package wolf.shin.earlyadopter.model.screen_router
+
+enum class AccountScreenRouter(val route: String) {
+    ROOT("account/"),
+    DETAIL("account/detail"),
+}
