--- presentation/src/main/java/wolf/shin/earlyadopter/model/screen_router/MoreScreenRouter.kt
@@ -1,5 +1,0 @@
-package wolf.shin.earlyadopter.model.screen_router
-enum class MoreScreenRouter(val route: String) {
-    ROOT("more/"),
-    DETAIL("more/detail"),
-}
